+++ include/mettle/driver/run_tests.hpp
@@ -66,7 +66,10 @@ namespace mettle {
         parents.push(suite.name());
 
         for(const auto &test : suite.tests()) {
-          const test_name name = {parents.all(), test.name, test.id};
+          const test_name name = {
+            test.id, parents.all(), test.name, test.location.file_name(),
+            test.location.line()
+          };
           auto action = filter(name, test.attrs);
           if(action.action == test_action::indeterminate)
             action = filter_by_attr(test.attrs);