+++ include/mettle/driver/test_name.hpp
@@ -1,6 +1,7 @@
 #ifndef INC_METTLE_DRIVER_TEST_NAME_HPP
 #define INC_METTLE_DRIVER_TEST_NAME_HPP
 
+#include <cstdint>
 #include <sstream>
 #include <string>
 #include <vector>
@@ -10,9 +11,10 @@
 namespace mettle {
 
   struct test_name {
-    std::vector<std::string> suites;
-    std::string name;
     test_uid id;
+    std::vector<std::string> suites;
+    std::string name, file;
+    std::uint_least32_t line = 0;
 
     std::string full_name() const {
       std::ostringstream ss;