+++ test/driver/log/test_counter.cpp
@@ -38,22 +38,23 @@ suite<logger_factory> test_verbose("counter logger", [](auto &_) {
   });
 
   _.test("started_test()", [](logger_factory &f) {
-    f.logger.started_test({{"suite"}, "test", 1});
+    f.logger.started_test({1, {"suite"}, "test", "file.cpp", 10});
     expect(f.ss.str(), equal_to(""));
   });
 
   _.test("passed_test()", [](logger_factory &f) {
-    f.logger.passed_test({{"suite"}, "test", 1}, {}, 0ms);
+    f.logger.passed_test({1, {"suite"}, "test", "file.cpp", 10}, {}, 0ms);
     expect(f.ss.str(), equal_to("\r[   1 |   1 |   0 |   0 ]"));
   });
 
   _.test("failed_test()", [](logger_factory &f) {
-    f.logger.failed_test({{"suite"}, "test", 1}, "error", {}, 0ms);
+    f.logger.failed_test({1, {"suite"}, "test", "file.cpp", 10}, "error", {},
+                         0ms);
     expect(f.ss.str(), equal_to("\r[   1 |   0 |   0 |   1 ]"));
   });
 
   _.test("skipped_test()", [](logger_factory &f) {
-    f.logger.skipped_test({{"suite"}, "test", 1}, "message");
+    f.logger.skipped_test({1, {"suite"}, "test", "file.cpp", 10}, "message");
     expect(f.ss.str(), equal_to("\r[   1 |   0 |   1 |   0 ]"));
   });